Added Role Upgrade Possibility
Added a way for members to become admins, here how it goes:
- You have to go in the Profile section of the user. It will display what kind of user it is.
- If just a member: there's a link to solve a riddle to become an admin.
- ANSWER: 22

If not able to answer properly: wrong answer displayed
If right answer, comes back to Profile page with updated status. The link should no longer appear.

For now, there is no way for an admin to become a member again.

To check if member in code, you can use $user->is_admin or $user->is_member. Please note that is_admin and is_member should be exclusive (can't be both).

This is related to #66 #68